requirements.txt
pandas
numpy
yfinance
pyyaml
ta
scikit-learn
torch
optuna  